+++ src/runtime/vm/appdomain.cpp
@@ -133,7 +133,7 @@ const char* AppDomain::get_friendly_name()
 
 RtObject* AppDomain::get_domain_data(RtString* name)
 {
-    utils::Utf16StrWithLen key(String::get_chars_ptr(name), String::get_length(name));
+    utils::Utf16StrWithLen key(String::get_chars_ptr(name), static_cast<size_t>(String::get_length(name)));
     auto it = g_appdomain_private_data.find(key);
     if (it != g_appdomain_private_data.end())
     {
@@ -144,7 +144,7 @@ RtObject* AppDomain::get_domain_data(RtString* name)
 
 void AppDomain::set_domain_data(RtString* name, RtObject* data)
 {
-    utils::Utf16StrWithLen key(String::get_chars_ptr(name), String::get_length(name));
+    utils::Utf16StrWithLen key(String::get_chars_ptr(name), static_cast<size_t>(String::get_length(name)));
     auto it = g_appdomain_private_data.find(key);
     ;
     if (it != g_appdomain_private_data.end())
@@ -153,7 +153,7 @@ void AppDomain::set_domain_data(RtString* name, RtObject* data)
         return;
     }
 
-    const Utf16Char* new_chars = utils::StringUtil::strdup_utf16_without_null_terminator(key.str, static_cast<int32_t>(key.length));
+    const Utf16Char* new_chars = utils::StringUtil::strdup_utf16_without_null_terminator(key.str, key.length);
     key.str = new_chars;
     g_appdomain_private_data.insert({key, data});
 }
